Create README.md
3- Create a 3D illusion for a profile picture

where a 25 Year old cute BOY in a black shirt

Sitting casually on a Wingback chair. Wearing

sneakers, a black cricket cap, and sunglasses,

he looks ahead. ahead. The background

features "SMART" in big and capital white fonts

on the dark grey wall. There should not be his

shadow, and there are wings to make it appear

as if he is an angel